refactor: ReviewNameField -> ReviewGroupDataField 이름 변경
