+++ frontend/src/pages/HomePage/components/Inputs/index.ts
@@ -1,4 +1,3 @@
-export { default as ProjectNameField } from './ProjectNameField';
-export { default as RevieweeNameField } from './RevieweeNameField';
+export { default as ReviewGroupDataField } from './ReviewGroupDataField';
 export { default as PasswordField } from './PasswordField';
 export { default as InputField } from './InputField';
